Refactor TabContext - RSC-293
