+++ lib/src/components/NxTabs/NxTab.tsx
@@ -7,7 +7,7 @@
 import React, { useContext } from 'react';
 import classnames from 'classnames';
 
-import { TabContext } from '../NxTabs/NxTabs';
+import TabContext from './TabContext';
 
 import { NxTabProps } from './types';
 export { NxTabProps };
